backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.27.2
beautifulsoup4==4.12.3
feedparser==6.0.11
playwright==1.49.0
apscheduler==3.10.4
ollama==0.4.4
python-multipart==0.0.12
python-dateutil==2.9.0
lxml==5.3.0
requests==2.32.3
pydantic==2.10.3
pydantic-settings==2.7.0
websockets==14.1
